scraping/requirements.txt
psycopg2-binary==2.9.3
python-dateutil==2.8.2
selenium==4.2.0
webdriver-manager
dotenv